refactor(contract): move FolderSummary* + Writer trait to workflow_orchestrator (brief §3.4, #44)
Spec architecture diagram places these in contract/workflow_orchestrator.rs;
implementation initially put them in domain/folder/mod.rs by accident.
Move per spec, including the trait-object-compiles test.
