+++ crates/cairn-core/src/contract/workflow_orchestrator.rs
@@ -3,7 +3,10 @@
 //! P0 default: tokio + `SQLite`-backed job table (#89). Optional Temporal
 //! adapter is a P1+ swap — same trait.
 
+use std::path::PathBuf;
+
 use crate::contract::version::{ContractVersion, VersionRange};
+use crate::domain::Rfc3339Timestamp;
 
 /// Contract version for `WorkflowOrchestrator`. Bumps when the trait surface changes.
 pub const CONTRACT_VERSION: ContractVersion = ContractVersion::new(0, 1, 0);
@@ -36,6 +39,53 @@ pub trait WorkflowOrchestrator: Send + Sync {
 
     /// Range of `WorkflowOrchestrator::CONTRACT_VERSION` values this impl accepts.
     fn supported_contract_versions(&self) -> VersionRange;
+}
+
+/// Schema for `_summary.md` (brief §3.4). P0 ships types only — body
+/// generation is the P1 `FolderSummaryWorkflow`.
+#[derive(Debug, Clone, PartialEq)]
+pub struct FolderSummary {
+    /// Vault-relative folder path.
+    pub folder: PathBuf,
+    /// When this summary was generated.
+    pub generated_at: Rfc3339Timestamp,
+    /// Agent that generated the summary (e.g. `agt:cairn-librarian:v2`).
+    pub generated_by: String,
+    /// Number of records the summary covers.
+    pub covers_records: u32,
+    /// Approximate token count of `body`.
+    pub summary_tokens: u32,
+    /// Generated prose body.
+    pub body: String,
+}
+
+/// Errors raised by [`FolderSummaryWriter`] implementations.
+#[derive(Debug, thiserror::Error)]
+#[non_exhaustive]
+pub enum FolderSummaryError {
+    /// No summary writer is registered (P0 default).
+    #[error("folder summary writer not registered")]
+    Unimplemented,
+    /// Internal error from the writer implementation.
+    #[error("folder summary writer internal: {0}")]
+    Internal(String),
+}
+
+/// Workflow-owned write surface for `_summary.md`. P0 ships zero
+/// implementations; P1 `cairn-workflows::FolderSummaryWorkflow` is the
+/// first implementor.
+#[async_trait::async_trait]
+pub trait FolderSummaryWriter: Send + Sync {
+    /// Persist a generated [`FolderSummary`] as `_summary.md` under its
+    /// folder, atomically. Implementors are responsible for I/O safety
+    /// (atomic rename, symlink rejection).
+    ///
+    /// # Errors
+    ///
+    /// Returns [`FolderSummaryError::Unimplemented`] when no writer is
+    /// registered, or [`FolderSummaryError::Internal`] for I/O / encoding
+    /// failures.
+    async fn write_summary(&self, summary: FolderSummary) -> Result<(), FolderSummaryError>;
 }
 
 /// Static identity descriptor for a [`WorkflowOrchestrator`] plugin (§4.1).
@@ -91,5 +141,20 @@ mod tests {
     fn static_consts_accessible() {
         assert_eq!(StubOrch::NAME, "stub-orch");
         assert!(StubOrch::SUPPORTED_VERSIONS.accepts(CONTRACT_VERSION));
+    }
+
+    #[test]
+    fn folder_summary_writer_trait_object_compiles() {
+        struct Stub;
+        #[async_trait::async_trait]
+        impl FolderSummaryWriter for Stub {
+            async fn write_summary(
+                &self,
+                _summary: FolderSummary,
+            ) -> Result<(), FolderSummaryError> {
+                Err(FolderSummaryError::Unimplemented)
+            }
+        }
+        let _: Box<dyn FolderSummaryWriter> = Box::new(Stub);
     }
 }
